removed View option from published files
